เกือบ

+++ module/deletecus.php
@@ -1,5 +1,9 @@
 <?php
+session_start();
+if (isset($_SESSION["valid_uname"]) && isset($_SESSION["valid_upass"]) && isset($_SESSION["valid_utype"])) {
     include 'connect.php';
+
+    
     $cus_id = $_GET['cus_id'];
 
     $sql = "DELETE FROM customer WHERE cus_id = '$cus_id'";
@@ -24,3 +28,9 @@
     
 </body>
 </html>
+<?php
+} else {
+    echo "<script> alert('Please Login'); window.location='frm_login.php';</script>";
+    exit();
+}
+?>
